only top member can do chats

+++ backend/src/module/chat-module/feature/room-member/exit-room-member/exit-room-listing.handler.ts
@@ -9,7 +9,22 @@ export class ExitRoomMemberService {
     ) { }
 
     async handle(body: RoomMemberDeletedMQEventPayload) {
+        const isMemberExists = await this.roomMemberRepository.findByUuid(body.room_member_uuid);
+        if (!isMemberExists) {
+            throw new BadRequestException("Member not found");
+        }
+
         await this.roomMemberRepository.deleteRoomMember(body.room_member_uuid);
+
+        if (isMemberExists.is_writer) {
+            const topMembers = await this.roomMemberRepository.findTopMembers(isMemberExists.room_uuid);
+            if (topMembers.length) {
+                for (const member of topMembers) {
+                    member.is_writer = true;
+                    await this.roomMemberRepository.save(member);
+                }
+            }
+        }
         return;
     }
 }
